+++ server/marketplace/src/main/java/com/dspajic/marketplace/service/LocationService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface LocationService {
     List<Location> getAllLocations();
-    Location getLocationById(Integer id);
-    Integer addLocation(Location entity);
+    Location getLocationById(Long id);
+    Long addLocation(Location entity);
     Integer updateLocation(Location entity);
     void deleteLocation(Integer id);
 }
